requirements.txt
streamlit
pymupdf
groq
python-dotenv
apify-client
